Ajout du delete  Please enter the commit message for your changes. Lines starting  with '' will be ignored, and an empty message aborts the commit.
 On branch main
 Your branch is up to date with 'origin/main'.

 Changes to be committed:
	modified:   service_info/db.sqlite3
	modified:   service_info/ticket/__pycache__/forms.cpython-312.pyc
	modified:   service_info/ticket/__pycache__/models.cpython-312.pyc
	modified:   service_info/ticket/__pycache__/urls.cpython-312.pyc
	modified:   service_info/ticket/__pycache__/views.cpython-312.pyc
	modified:   service_info/ticket/forms.py
	modified:   service_info/ticket/models.py
	modified:   service_info/ticket/static/style.css
	modified:   service_info/ticket/templates/ticket/index.html
	modified:   service_info/ticket/templates/ticket/main.html
	modified:   service_info/ticket/urls.py
	modified:   service_info/ticket/views.py

+++ service_info/ticket/views.py
@@ -1,14 +1,11 @@
 from django.shortcuts import render, redirect
-
-from .models import Ticket
-from .forms import *
-
+from .models import Ticket, Category
+from .forms import TicketForm, CategoryForm, TicketIdForm
 
 def index(request):
     tickets = Ticket.objects.all()
     return render(request, 'ticket/index.html', {'tickets': tickets})
 
-## CRUD pour les tickets
 def create_ticket(request):
     if request.method == 'POST':
         form = TicketForm(request.POST)
@@ -20,14 +17,16 @@ def create_ticket(request):
     return render(request, 'ticket/create_ticket.html', {'form': form})
 
 def read_ticket(request, ticket_id):
-    ticket = Ticket.objects.get(id=ticket_id)
-    return render(request, 'ticket/affiche_ticket.html', {'ticket': ticket})
+    if request.method == 'GET':
+        ticket = Ticket.objects.get(id=ticket_id)
+        return render(request, 'ticket/affiche_ticket.html', {'ticket': ticket})
 
 def delete_ticket(request, ticket_id):
-    ticket = Ticket.objects.get(id=ticket_id)
-    ticket.delete()
-    return redirect('index')
-
+    if request.method == 'GET':
+        ticket = Ticket.objects.get(id=ticket_id)
+        ticket.delete()
+        return redirect('ticket:index')
+    
 def update_ticket(request, ticket_id):
     ticket = Ticket.objects.get(id=ticket_id)
     form = TicketForm(instance=ticket)
@@ -42,27 +41,29 @@ def traitement_ticket(request):
     lform = TicketForm(request.POST)
     if lform.is_valid():
         ticket = lform.save()
-        return render(request, 'ticket/affiche_ticket.html', {'Ticket': ticket})
+        return render(request, 'ticket/affiche_ticket.html', {'ticket': ticket})
     else:
         return render(request, 'ticket/create_ticket.html', {'form': lform})
-    
     
 def search_ticket(request):
     if request.method == 'POST':
         lform = TicketIdForm(request.POST)
         if lform.is_valid():
             ticket_id = lform.cleaned_data['ticket_id']
-            return redirect('ticket:read_ticket', ticket_id=ticket_id)
+            action = lform.cleaned_data['action']
+            
+            if action == 'read_ticket':
+                return redirect('ticket:read_ticket', ticket_id=ticket_id)
+            elif action == 'delete_ticket':
+                return redirect('ticket:delete_ticket', ticket_id=ticket_id)
+            elif action == 'update_ticket':
+                return redirect('ticket:update_ticket', ticket_id=ticket_id)
+            
     else:
         lform = TicketIdForm()
     return render(request, 'ticket/search_ticket.html', {'form': lform})
-    
-    
-    
-    
-    
+
 def create_category(request):
-    """Création d'une catégorie."""
     if request.method == 'POST':
         form = CategoryForm(request.POST)
         if form.is_valid():
@@ -83,18 +84,15 @@ def traitement_category(request):
     return render(request, 'ticket/create_category.html', {'form': form})
 
 def read_category(request, category_id):
-    """Recherche par id de la catégorie et affiche les détails de la catégorie."""
     category = Category.objects.get(id=category_id)
     return render(request, 'ticket/affiche_category.html', {'category': category})
 
 def delete_category(request, category_id):
-    """Suppression d'une catégorie par id."""
     category = Category.objects.get(id=category_id)
     category.delete()
     return redirect('index')
 
 def update_category(request, category_id):
-    """Update d'une catégorie par id."""
     category = Category.objects.get(id=category_id)
     form = CategoryForm(instance=category)
     if request.method == 'POST':
